Added more comments.
P.S. I was having trouble getting the external class on a different file working with the Program.cs file, so I could only create a finished and working product with both classes on the one file. I didn't see the criteria about there having to be the separate class file until I got home late the night this assignment was due.